Extract messages as a module
