+++ tests/test_deepanimebot_messages.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
-
 import deploy
-from deepanimebot import bot
+from deepanimebot import messages
 
 
 def test_my_guess_honors_max_length_by_truncating_longest():
@@ -12,7 +11,7 @@ def test_my_guess_honors_max_length_by_truncating_longest():
     # before truncation:
     # '\n2. 567890 2.40%\n7. 012 4.60%'
     max_length = 26
-    reply = bot.Messages.my_guess(y, preface='', max_length=max_length)
+    reply = messages.Messages.my_guess(y, preface='', max_length=max_length)
     assert reply == '\n2. 567 2.40%\n7. 012 4.60%'
     assert len(reply) <= max_length
 
@@ -25,6 +24,6 @@ def test_my_guess_honors_max_length_by_truncating_all():
     # before truncation:
     # '\n2. 567 9.10%\n4. 789 1.20%'
     max_length = 22
-    reply = bot.Messages.my_guess(y, preface='', max_length=max_length)
+    reply = messages.Messages.my_guess(y, preface='', max_length=max_length)
     assert reply == '\n2. 5 9.10%\n4. 7 1.20%'
     assert len(reply) <= max_length
